wasm/sdk: move existing benchmarks to new file
